Simulate data (utilized for function tests)

Simulate data (utilized for function tests)

Usage

gensurv(seed, n1, n2, obsv_duration, lambda1, lambda2, unit = "Months")

Arguments

seed

A numerical input or object that will set the standard for data simulation.

n1

A numerical input or object that represents the desired total population for the simulated data (for treatment group 1).

n2

A numerical input or object that represents the desired total population for the simulated data (for treatment group 2).

obsv_duration

A numerical input or object that is the maximum time the observational period will span for the simulated data.

lambda1

A numerical input or object that is the first lambda that will be used in the data simulation (used as the first parameter for simulation).

lambda2

A numerical input or object that is the second lambda that will be used in the data simulation (used as the second parameter for simulation).

unit

A non-numerical input that specifies the length of the observational period ("Week", "Month", etc.).

Value

A simulated dataframe.

Examples

gensurv(111, 2000, 1000, 36, .005, .0048, "Weeks")
#>       eventtime    diff obsv_duration obsv_unit
#> 1    0.01043160  0.0000            36     Weeks
#> 2    0.09792045  0.0005            36     Weeks
#> 3    0.22643677  0.0010            36     Weeks
#> 4    0.61189349  0.0000            36     Weeks
#> 5    0.74987781  0.0005            36     Weeks
#> 6    0.97980675  0.0010            36     Weeks
#> 7    0.98337377  0.0015            36     Weeks
#> 8    0.98476594  0.0005            36     Weeks
#> 9    1.06245322  0.0010            36     Weeks
#> 10   1.16755774  0.0015            36     Weeks
#> 11   1.17658630  0.0020            36     Weeks
#> 12   1.20696182  0.0010            36     Weeks
#> 13   1.27975392  0.0000            36     Weeks
#> 14   1.29162677 -0.0010            36     Weeks
#> 15   1.29721199 -0.0005            36     Weeks
#> 16   1.34867156  0.0000            36     Weeks
#> 17   1.36273360  0.0005            36     Weeks
#> 18   1.39395281  0.0010            36     Weeks
#> 19   1.42064513  0.0015            36     Weeks
#> 20   1.47629083  0.0020            36     Weeks
#> 21   1.47783986  0.0025            36     Weeks
#> 22   1.48142663  0.0015            36     Weeks
#> 23   1.54596237  0.0005            36     Weeks
#> 24   1.62650418 -0.0005            36     Weeks
#> 25   1.63115749 -0.0015            36     Weeks
#> 26   1.68292485 -0.0010            36     Weeks
#> 27   1.85104930 -0.0005            36     Weeks
#> 28   1.91704507 -0.0015            36     Weeks
#> 29   1.97047906 -0.0010            36     Weeks
#> 30   2.09100946 -0.0005            36     Weeks
#> 31   2.26721597  0.0000            36     Weeks
#> 32   2.27213574  0.0005            36     Weeks
#> 33   2.34049754  0.0010            36     Weeks
#> 34   2.34067301  0.0000            36     Weeks
#> 35   2.49351806 -0.0010            36     Weeks
#> 36   2.50857133 -0.0005            36     Weeks
#> 37   2.56311151 -0.0015            36     Weeks
#> 38   2.58121804 -0.0010            36     Weeks
#> 39   2.63385985 -0.0005            36     Weeks
#> 40   2.68631762  0.0000            36     Weeks
#> 41   2.76439790  0.0005            36     Weeks
#> 42   2.82226439 -0.0005            36     Weeks
#> 43   2.83064598 -0.0015            36     Weeks
#> 44   2.96080529 -0.0010            36     Weeks
#> 45   2.97939768 -0.0020            36     Weeks
#> 46   3.00863179 -0.0030            36     Weeks
#> 47   3.04721467 -0.0025            36     Weeks
#> 48   3.08686118 -0.0035            36     Weeks
#> 49   3.09444962 -0.0030            36     Weeks
#> 50   3.09922545 -0.0040            36     Weeks
#> 51   3.31454504 -0.0035            36     Weeks
#> 52   3.36092841 -0.0030            36     Weeks
#> 53   3.47788815 -0.0040            36     Weeks
#> 54   3.48761484 -0.0050            36     Weeks
#> 55   3.50991199 -0.0045            36     Weeks
#> 56   3.79327899 -0.0040            36     Weeks
#> 57   4.04934535 -0.0050            36     Weeks
#> 58   4.12652805 -0.0060            36     Weeks
#> 59   4.12683015 -0.0055            36     Weeks
#> 60   4.12711665 -0.0050            36     Weeks
#> 61   4.13502322 -0.0060            36     Weeks
#> 62   4.31222336 -0.0055            36     Weeks
#> 63   4.33909737 -0.0050            36     Weeks
#> 64   4.38365893 -0.0045            36     Weeks
#> 65   4.45145808 -0.0040            36     Weeks
#> 66   4.46494438 -0.0035            36     Weeks
#> 67   4.53363319 -0.0045            36     Weeks
#> 68   4.65649638 -0.0040            36     Weeks
#> 69   4.74573649 -0.0050            36     Weeks
#> 70   4.77676710 -0.0045            36     Weeks
#> 71   4.84702527 -0.0055            36     Weeks
#> 72   5.10784948 -0.0050            36     Weeks
#> 73   5.13039471 -0.0045            36     Weeks
#> 74   5.19336458 -0.0055            36     Weeks
#> 75   5.24640728 -0.0050            36     Weeks
#> 76   5.40620518 -0.0060            36     Weeks
#> 77   5.45405136 -0.0055            36     Weeks
#> 78   5.45588539 -0.0065            36     Weeks
#> 79   5.49358386 -0.0060            36     Weeks
#> 80   5.60943698 -0.0055            36     Weeks
#> 81   5.61259291 -0.0065            36     Weeks
#> 82   5.68558979 -0.0060            36     Weeks
#> 83   5.76882847 -0.0055            36     Weeks
#> 84   5.87025523 -0.0050            36     Weeks
#> 85   6.00944531 -0.0045            36     Weeks
#> 86   6.06241613 -0.0040            36     Weeks
#> 87   6.06294695 -0.0035            36     Weeks
#> 88   6.09321319 -0.0030            36     Weeks
#> 89   6.16519991 -0.0025            36     Weeks
#> 90   6.20558380 -0.0035            36     Weeks
#> 91   6.22309281 -0.0045            36     Weeks
#> 92   6.33692109 -0.0055            36     Weeks
#> 93   6.49229266 -0.0050            36     Weeks
#> 94   6.75364829 -0.0045            36     Weeks
#> 95   6.76089419 -0.0040            36     Weeks
#> 96   7.00902994 -0.0035            36     Weeks
#> 97   7.02310928 -0.0030            36     Weeks
#> 98   7.11931867 -0.0040            36     Weeks
#> 99   7.38245443 -0.0050            36     Weeks
#> 100  7.42317043 -0.0045            36     Weeks
#> 101  7.48195096 -0.0040            36     Weeks
#> 102  7.49901374 -0.0035            36     Weeks
#> 103  7.55204776 -0.0030            36     Weeks
#> 104  7.82205610 -0.0025            36     Weeks
#> 105  7.84439742 -0.0035            36     Weeks
#> 106  7.94315718 -0.0030            36     Weeks
#> 107  7.95081991 -0.0040            36     Weeks
#> 108  8.00736227 -0.0035            36     Weeks
#> 109  8.02150359 -0.0030            36     Weeks
#> 110  8.08809314 -0.0040            36     Weeks
#> 111  8.11991625 -0.0035            36     Weeks
#> 112  8.14932782 -0.0045            36     Weeks
#> 113  8.17266360 -0.0040            36     Weeks
#> 114  8.24204981 -0.0035            36     Weeks
#> 115  8.25677339 -0.0030            36     Weeks
#> 116  8.45286617 -0.0025            36     Weeks
#> 117  8.52827226 -0.0020            36     Weeks
#> 118  8.54049605 -0.0015            36     Weeks
#> 119  8.57503445 -0.0010            36     Weeks
#> 120  8.64093755 -0.0005            36     Weeks
#> 121  8.68216002  0.0000            36     Weeks
#> 122  8.76047895  0.0005            36     Weeks
#> 123  8.90777472  0.0010            36     Weeks
#> 124  8.96792234  0.0015            36     Weeks
#> 125  9.07953546  0.0005            36     Weeks
#> 126  9.08655351  0.0010            36     Weeks
#> 127  9.17040461  0.0015            36     Weeks
#> 128  9.19271956  0.0005            36     Weeks
#> 129  9.21240685 -0.0005            36     Weeks
#> 130  9.33652006 -0.0015            36     Weeks
#> 131  9.34831911 -0.0010            36     Weeks
#> 132  9.35176589 -0.0005            36     Weeks
#> 133  9.60107455 -0.0015            36     Weeks
#> 134  9.63877775 -0.0010            36     Weeks
#> 135  9.84546644 -0.0005            36     Weeks
#> 136  9.85907611 -0.0015            36     Weeks
#> 137  9.87758205 -0.0025            36     Weeks
#> 138  9.95770434 -0.0020            36     Weeks
#> 139 10.12311772 -0.0015            36     Weeks
#> 140 10.18245956 -0.0025            36     Weeks
#> 141 10.26069819 -0.0035            36     Weeks
#> 142 10.27248308 -0.0045            36     Weeks
#> 143 10.36461582 -0.0040            36     Weeks
#> 144 10.45187398 -0.0035            36     Weeks
#> 145 10.52696768 -0.0030            36     Weeks
#> 146 10.53030674 -0.0025            36     Weeks
#> 147 10.57670871 -0.0035            36     Weeks
#> 148 10.62545281 -0.0030            36     Weeks
#> 149 10.62954494 -0.0025            36     Weeks
#> 150 10.69352222 -0.0020            36     Weeks
#> 151 10.70875163 -0.0030            36     Weeks
#> 152 10.71263779 -0.0025            36     Weeks
#> 153 10.73492371 -0.0020            36     Weeks
#> 154 10.74251126 -0.0015            36     Weeks
#> 155 10.74489950 -0.0025            36     Weeks
#> 156 10.77040223 -0.0020            36     Weeks
#> 157 10.79345820 -0.0015            36     Weeks
#> 158 10.98786802 -0.0010            36     Weeks
#> 159 11.26276814 -0.0005            36     Weeks
#> 160 11.39100666  0.0000            36     Weeks
#> 161 11.40427262  0.0005            36     Weeks
#> 162 11.40930985 -0.0005            36     Weeks
#> 163 11.42063902  0.0000            36     Weeks
#> 164 11.45067335  0.0005            36     Weeks
#> 165 11.60749346  0.0010            36     Weeks
#> 166 11.73351583  0.0015            36     Weeks
#> 167 11.78095631  0.0020            36     Weeks
#> 168 11.82917623  0.0025            36     Weeks
#> 169 12.22235660  0.0030            36     Weeks
#> 170 12.27658758  0.0035            36     Weeks
#> 171 12.28153137  0.0040            36     Weeks
#> 172 12.33014003  0.0045            36     Weeks
#> 173 12.36806207  0.0050            36     Weeks
#> 174 12.41308029  0.0040            36     Weeks
#> 175 12.53893038  0.0030            36     Weeks
#> 176 12.59064781  0.0020            36     Weeks
#> 177 12.61975040  0.0025            36     Weeks
#> 178 12.67063454  0.0015            36     Weeks
#> 179 12.92375327  0.0005            36     Weeks
#> 180 12.98995871  0.0010            36     Weeks
#> 181 13.01655360  0.0015            36     Weeks
#> 182 13.09549474  0.0020            36     Weeks
#> 183 13.32683075  0.0025            36     Weeks
#> 184 13.33859954  0.0030            36     Weeks
#> 185 13.49689561  0.0035            36     Weeks
#> 186 13.50199266  0.0040            36     Weeks
#> 187 13.79634437  0.0045            36     Weeks
#> 188 13.84702837  0.0050            36     Weeks
#> 189 13.96757667  0.0040            36     Weeks
#> 190 14.18705161  0.0030            36     Weeks
#> 191 14.22890032  0.0035            36     Weeks
#> 192 14.27548762  0.0040            36     Weeks
#> 193 14.31084844  0.0030            36     Weeks
#> 194 14.58322355  0.0035            36     Weeks
#> 195 14.59929977  0.0040            36     Weeks
#> 196 14.79168287  0.0045            36     Weeks
#> 197 14.87023580  0.0035            36     Weeks
#> 198 14.94873738  0.0040            36     Weeks
#> 199 15.03544157  0.0045            36     Weeks
#> 200 15.05648647  0.0050            36     Weeks
#> 201 15.10930606  0.0055            36     Weeks
#> 202 15.16106731  0.0060            36     Weeks
#> 203 15.37836943  0.0050            36     Weeks
#> 204 15.51253663  0.0055            36     Weeks
#> 205 15.58337906  0.0045            36     Weeks
#> 206 15.59332378  0.0050            36     Weeks
#> 207 15.62157294  0.0055            36     Weeks
#> 208 15.74275734  0.0045            36     Weeks
#> 209 15.82311390  0.0050            36     Weeks
#> 210 15.84495972  0.0055            36     Weeks
#> 211 15.88153705  0.0060            36     Weeks
#> 212 15.93918254  0.0050            36     Weeks
#> 213 16.21331384  0.0055            36     Weeks
#> 214 16.28652558  0.0045            36     Weeks
#> 215 16.29260511  0.0050            36     Weeks
#> 216 16.39287384  0.0055            36     Weeks
#> 217 16.47532676  0.0060            36     Weeks
#> 218 16.64805765  0.0065            36     Weeks
#> 219 16.73470532  0.0070            36     Weeks
#> 220 16.77972144  0.0075            36     Weeks
#> 221 16.78730794  0.0080            36     Weeks
#> 222 16.93273799  0.0085            36     Weeks
#> 223 16.93564655  0.0090            36     Weeks
#> 224 17.14604027  0.0080            36     Weeks
#> 225 17.28003977  0.0085            36     Weeks
#> 226 17.33229942  0.0090            36     Weeks
#> 227 17.35053078  0.0080            36     Weeks
#> 228 17.35595470  0.0085            36     Weeks
#> 229 17.45972143  0.0075            36     Weeks
#> 230 17.52024387  0.0065            36     Weeks
#> 231 17.53111025  0.0055            36     Weeks
#> 232 17.53741175  0.0060            36     Weeks
#> 233 17.54065868  0.0065            36     Weeks
#> 234 17.54300329  0.0070            36     Weeks
#> 235 17.76284509  0.0075            36     Weeks
#> 236 17.87260576  0.0065            36     Weeks
#> 237 17.90678652  0.0070            36     Weeks
#> 238 17.94813792  0.0060            36     Weeks
#> 239 17.95947632  0.0050            36     Weeks
#> 240 17.99004915  0.0055            36     Weeks
#> 241 18.03622929  0.0060            36     Weeks
#> 242 18.05164965  0.0065            36     Weeks
#> 243 18.15675829  0.0055            36     Weeks
#> 244 18.25998499  0.0045            36     Weeks
#> 245 18.28326343  0.0035            36     Weeks
#> 246 18.40107283  0.0025            36     Weeks
#> 247 18.52856881  0.0030            36     Weeks
#> 248 18.66026563  0.0035            36     Weeks
#> 249 18.73694246  0.0040            36     Weeks
#> 250 18.77025196  0.0045            36     Weeks
#> 251 19.14103982  0.0050            36     Weeks
#> 252 19.17039242  0.0040            36     Weeks
#> 253 19.18461699  0.0045            36     Weeks
#> 254 19.23289269  0.0050            36     Weeks
#> 255 19.25767580  0.0055            36     Weeks
#> 256 19.36774064  0.0060            36     Weeks
#> 257 19.46411622  0.0065            36     Weeks
#> 258 19.52951567  0.0055            36     Weeks
#> 259 19.72850090  0.0045            36     Weeks
#> 260 19.78417238  0.0050            36     Weeks
#> 261 19.82487624  0.0040            36     Weeks
#> 262 19.90078853  0.0045            36     Weeks
#> 263 19.93494455  0.0035            36     Weeks
#> 264 19.98247654  0.0040            36     Weeks
#> 265 19.98254877  0.0045            36     Weeks
#> 266 19.99025539  0.0035            36     Weeks
#> 267 20.14891058  0.0040            36     Weeks
#> 268 20.16728415  0.0045            36     Weeks
#> 269 20.24409693  0.0050            36     Weeks
#> 270 20.26852187  0.0055            36     Weeks
#> 271 20.32353155  0.0060            36     Weeks
#> 272 20.38112231  0.0065            36     Weeks
#> 273 20.46646166  0.0070            36     Weeks
#> 274 20.53683493  0.0075            36     Weeks
#> 275 20.56999403  0.0080            36     Weeks
#> 276 20.66222052  0.0085            36     Weeks
#> 277 20.82225185  0.0075            36     Weeks
#> 278 20.87835129  0.0065            36     Weeks
#> 279 20.88232092  0.0055            36     Weeks
#> 280 21.21335337  0.0060            36     Weeks
#> 281 21.26663656  0.0065            36     Weeks
#> 282 21.36981762  0.0055            36     Weeks
#> 283 21.86764955  0.0060            36     Weeks
#> 284 21.94714541  0.0065            36     Weeks
#> 285 21.97657972  0.0070            36     Weeks
#> 286 22.02093929  0.0075            36     Weeks
#> 287 22.03054236  0.0080            36     Weeks
#> 288 22.05844533  0.0085            36     Weeks
#> 289 22.08483496  0.0090            36     Weeks
#> 290 22.15638838  0.0095            36     Weeks
#> 291 22.32958285  0.0100            36     Weeks
#> 292 22.46328697  0.0105            36     Weeks
#> 293 22.99440212  0.0095            36     Weeks
#> 294 23.05396392  0.0085            36     Weeks
#> 295 23.15916732  0.0090            36     Weeks
#> 296 23.21338007  0.0080            36     Weeks
#> 297 23.36703264  0.0085            36     Weeks
#> 298 23.49998142  0.0090            36     Weeks
#> 299 23.59791000  0.0080            36     Weeks
#> 300 23.68724673  0.0085            36     Weeks
#> 301 23.76267098  0.0090            36     Weeks
#> 302 23.82085477  0.0080            36     Weeks
#> 303 23.93242053  0.0085            36     Weeks
#> 304 23.96165905  0.0075            36     Weeks
#> 305 23.99273206  0.0065            36     Weeks
#> 306 24.02270025  0.0070            36     Weeks
#> 307 24.05462880  0.0060            36     Weeks
#> 308 24.11250396  0.0065            36     Weeks
#> 309 24.19525741  0.0070            36     Weeks
#> 310 24.23440626  0.0075            36     Weeks
#> 311 24.25913938  0.0080            36     Weeks
#> 312 24.30211979  0.0070            36     Weeks
#> 313 24.32977537  0.0075            36     Weeks
#> 314 24.73506225  0.0065            36     Weeks
#> 315 24.89048946  0.0070            36     Weeks
#> 316 25.00323274  0.0060            36     Weeks
#> 317 25.14372769  0.0050            36     Weeks
#> 318 25.24102770  0.0040            36     Weeks
#> 319 25.24550598  0.0045            36     Weeks
#> 320 25.30206534  0.0050            36     Weeks
#> 321 25.32453650  0.0055            36     Weeks
#> 322 25.50432856  0.0060            36     Weeks
#> 323 25.54783525  0.0065            36     Weeks
#> 324 25.59779352  0.0070            36     Weeks
#> 325 25.82961703  0.0060            36     Weeks
#> 326 25.87222579  0.0050            36     Weeks
#> 327 25.87951242  0.0040            36     Weeks
#> 328 25.89004559  0.0045            36     Weeks
#> 329 25.91866325  0.0035            36     Weeks
#> 330 25.93707782  0.0040            36     Weeks
#> 331 25.94108992  0.0045            36     Weeks
#> 332 26.31940704  0.0050            36     Weeks
#> 333 26.39173556  0.0055            36     Weeks
#> 334 26.40026685  0.0060            36     Weeks
#> 335 26.48643054  0.0065            36     Weeks
#> 336 26.51734130  0.0070            36     Weeks
#> 337 26.52943227  0.0075            36     Weeks
#> 338 26.55190647  0.0080            36     Weeks
#> 339 26.55321209  0.0085            36     Weeks
#> 340 26.56850168  0.0090            36     Weeks
#> 341 26.65389317  0.0080            36     Weeks
#> 342 26.65493130  0.0085            36     Weeks
#> 343 26.65885530  0.0090            36     Weeks
#> 344 26.71971950  0.0095            36     Weeks
#> 345 26.78133905  0.0100            36     Weeks
#> 346 26.79803404  0.0090            36     Weeks
#> 347 26.84372094  0.0095            36     Weeks
#> 348 26.94868493  0.0100            36     Weeks
#> 349 27.03406285  0.0105            36     Weeks
#> 350 27.12039678  0.0110            36     Weeks
#> 351 27.14579896  0.0115            36     Weeks
#> 352 27.23575104  0.0105            36     Weeks
#> 353 27.46512590  0.0110            36     Weeks
#> 354 27.51775189  0.0100            36     Weeks
#> 355 27.53987795  0.0105            36     Weeks
#> 356 27.57085394  0.0110            36     Weeks
#> 357 27.65863424  0.0115            36     Weeks
#> 358 27.68331186  0.0105            36     Weeks
#> 359 27.94288732  0.0110            36     Weeks
#> 360 28.07332179  0.0115            36     Weeks
#> 361 28.21773477  0.0120            36     Weeks
#> 362 28.25821871  0.0125            36     Weeks
#> 363 28.29786279  0.0130            36     Weeks
#> 364 28.36609730  0.0135            36     Weeks
#> 365 28.38908720  0.0125            36     Weeks
#> 366 28.40440568  0.0130            36     Weeks
#> 367 28.43393075  0.0135            36     Weeks
#> 368 28.47211237  0.0140            36     Weeks
#> 369 28.56632130  0.0130            36     Weeks
#> 370 28.64729805  0.0120            36     Weeks
#> 371 28.68276764  0.0125            36     Weeks
#> 372 28.76970441  0.0130            36     Weeks
#> 373 28.83989718  0.0135            36     Weeks
#> 374 28.91585551  0.0140            36     Weeks
#> 375 28.93700963  0.0145            36     Weeks
#> 376 29.08956684  0.0150            36     Weeks
#> 377 29.60635724  0.0155            36     Weeks
#> 378 29.65166042  0.0145            36     Weeks
#> 379 29.71306238  0.0150            36     Weeks
#> 380 29.74490478  0.0140            36     Weeks
#> 381 29.82575626  0.0145            36     Weeks
#> 382 29.89371536  0.0135            36     Weeks
#> 383 30.12632626  0.0140            36     Weeks
#> 384 30.21003179  0.0145            36     Weeks
#> 385 30.28870845  0.0150            36     Weeks
#> 386 30.29766298  0.0155            36     Weeks
#> 387 30.35825763  0.0160            36     Weeks
#> 388 30.49727416  0.0165            36     Weeks
#> 389 30.57223941  0.0170            36     Weeks
#> 390 30.58375007  0.0175            36     Weeks
#> 391 30.61569607  0.0180            36     Weeks
#> 392 30.72362998  0.0185            36     Weeks
#> 393 30.72894213  0.0175            36     Weeks
#> 394 30.77067056  0.0180            36     Weeks
#> 395 30.92264082  0.0185            36     Weeks
#> 396 30.95446053  0.0190            36     Weeks
#> 397 30.99389011  0.0195            36     Weeks
#> 398 31.08811549  0.0185            36     Weeks
#> 399 31.24839217  0.0190            36     Weeks
#> 400 31.33744891  0.0195            36     Weeks
#> 401 31.45217366  0.0200            36     Weeks
#> 402 31.51427915  0.0205            36     Weeks
#> 403 31.58373211  0.0210            36     Weeks
#> 404 31.63596134  0.0215            36     Weeks
#> 405 31.66461001  0.0205            36     Weeks
#> 406 31.66490090  0.0210            36     Weeks
#> 407 31.67634201  0.0215            36     Weeks
#> 408 31.73133947  0.0205            36     Weeks
#> 409 31.82256860  0.0210            36     Weeks
#> 410 31.84656041  0.0215            36     Weeks
#> 411 32.07927243  0.0205            36     Weeks
#> 412 32.10732129  0.0210            36     Weeks
#> 413 32.19694860  0.0215            36     Weeks
#> 414 32.20350091  0.0220            36     Weeks
#> 415 32.22521967  0.0225            36     Weeks
#> 416 32.31032752  0.0230            36     Weeks
#> 417 32.36963492  0.0220            36     Weeks
#> 418 32.37189871  0.0210            36     Weeks
#> 419 32.45862966  0.0215            36     Weeks
#> 420 32.61884909  0.0220            36     Weeks
#> 421 32.62217996  0.0225            36     Weeks
#> 422 32.77258363  0.0215            36     Weeks
#> 423 32.85154971  0.0220            36     Weeks
#> 424 32.86710292  0.0225            36     Weeks
#> 425 32.87472924  0.0230            36     Weeks
#> 426 32.91137542  0.0220            36     Weeks
#> 427 32.94347278  0.0225            36     Weeks
#> 428 33.00279897  0.0215            36     Weeks
#> 429 33.02365074  0.0205            36     Weeks
#> 430 33.19789003  0.0195            36     Weeks
#> 431 33.20265000  0.0185            36     Weeks
#> 432 33.24396762  0.0175            36     Weeks
#> 433 33.29721475  0.0180            36     Weeks
#> 434 33.32387214  0.0185            36     Weeks
#> 435 33.40077286  0.0175            36     Weeks
#> 436 33.40108957  0.0165            36     Weeks
#> 437 33.56102611  0.0170            36     Weeks
#> 438 33.58969390  0.0175            36     Weeks
#> 439 33.63813018  0.0165            36     Weeks
#> 440 33.72095672  0.0155            36     Weeks
#> 441 33.74886254  0.0160            36     Weeks
#> 442 33.76857254  0.0165            36     Weeks
#> 443 33.77163536  0.0155            36     Weeks
#> 444 33.80512544  0.0160            36     Weeks
#> 445 33.95512967  0.0165            36     Weeks
#> 446 34.00082622  0.0155            36     Weeks
#> 447 34.12435410  0.0160            36     Weeks
#> 448 34.14721507  0.0165            36     Weeks
#> 449 34.16961296  0.0170            36     Weeks
#> 450 34.22526525  0.0175            36     Weeks
#> 451 34.28546262  0.0180            36     Weeks
#> 452 34.42452297  0.0185            36     Weeks
#> 453 34.47647526  0.0190            36     Weeks
#> 454 34.52861796  0.0195            36     Weeks
#> 455 35.26474019  0.0200            36     Weeks
#> 456 35.30622276  0.0205            36     Weeks
#> 457 35.52925648  0.0210            36     Weeks
#> 458 35.55131424  0.0215            36     Weeks
#> 459 35.62517798  0.0220            36     Weeks
#> 460 35.65989026  0.0210            36     Weeks
#> 461 35.66904767  0.0215            36     Weeks
#> 462 35.72416991  0.0205            36     Weeks
#> 463 35.94805537  0.0195            36     Weeks
#> 464 36.00000000  0.0185            36     Weeks